READ ME
That's my first and only app on android. It's a very simple app in order to facilitate the way we send messages to our contacts. First, you have to save some messages and when you want to send one of these, to your contacts, you just select them and press the people from your contact list.